# Clinovation_ApiSpec.yaml
openapi: 3.0.3
info:
  title: Clinovation API
  description: API for Hospital Management
  version: 1.0.0
servers:
  - url: "http://localhost:8080/api/v1/"
  - url: "http://clinovation/api/v1/"
...
paths:
...
  /medical-staff/login:
    post:
      tags:
        - Medical Staff
      summary: Medical Staff Login
      operationId: loginMedicalStaff
      parameters:
        - name: email
          in: query
          description: The email medical staff for login
          required: true
          schema:
            type: string
        - name: password
          in: query
          description: The password medical staff for login
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: "#/components/schemas/Login"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FailResponse400"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FailResponse500"
...
  /doctor/login:
    post:
      tags:
        - Doctor
      summary: Doctor Login
      operationId: loginDoctor
      parameters:
        - name: email
          in: query
          description: The email doctor for login
          required: true
          schema:
            type: string
        - name: password
          in: query
          description: The password doctor for login
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: "#/components/schemas/Login"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FailResponse400"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FailResponse500"
  /doctor/register:
    post:
      tags:
        - Doctor
      summary: Register new Doctor
      parameters:
        - name: name
          in: query
          description: doctor Name
          required: true
          schema:
            type: string
        - name: nik
          in: query
          description: doctor Nik
          required: true
          schema:
            type: string
        - name: password
          in: query
          description: doctor Password
          required: true
          schema:
            type: string
        - name: dob
          in: query
          description: doctor Date of Birth
          required: true
          schema:
            type: string
        - name: sex
          in: query
          description: doctor Sex
          required: true
          schema:
            type: string
        - name: email
          in: query
          description: doctor Email
          required: true
          schema:
            type: string
        - name: contact
          in: query
          description: doctor Contact
          required: true
          schema:
            type: string
        - name: specialist
          in: query
          description: doctor specialist
          required: true
          schema:
            type: string
        - name: wokrk_experience
          in: query
          description: doctor wokrk_experience
          required: true
          schema:
            type: string
        - name: schedule
          in: query
          description: doctor schedule
          required: true
          schema:
            type: string
        - name: work_hour
          in: query
          description: doctor work_hour
          required: true
          schema:
            type: string

      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                properties:
                  meta:
                    $ref: "#/components/schemas/SuccessResponse"
                  data:
                    $ref: "#/components/schemas/DoctorResponses"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FailResponse400"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FailResponse500"
...
  /doctor:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - Doctor
      summary: Get all Doctor
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                properties:
                  meta:
                    $ref: "#/components/schemas/SuccessResponse"
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/DoctorResponses"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FailResponse404"
    delete:
      security:
        - bearerAuth: [ ]
      tags:
        - Doctor
      summary: Delete a specified doctor
      parameters:
        - in: path
          name: uuid
          schema:
            type: string
          description: uuid of doctor to be deleted
          required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                properties:
                  meta:
                    $ref: "#/components/schemas/SuccessResponse"
                  data:
                    type: string
                    example: doctor was Deleted
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FailResponse400"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FailResponse403"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FailResponse500"
    put:
      security:
        - bearerAuth: [ ]
      tags:
        - Doctor
      summary: Update doctor
      parameters:
        - name: name
          in: query
          description: doctor Name
          required: true
          schema:
            type: string
        - name: password
          in: query
          description: doctor Password
          required: true
          schema:
            type: string
        - name: dob
          in: query
          description: doctor Date of Birth
          required: true
          schema:
            type: string
        - name: sex
          in: query
          description: doctor Sex
          required: true
          schema:
            type: string
        - name: email
          in: query
          description: doctor Email
          required: true
          schema:
            type: string
        - name: contact
          in: query
          description: doctor Contact
          required: true
          schema:
            type: string
        - name: specialist
          in: query
          description: doctor specialist
          required: true
          schema:
            type: string
        - name: wokrk_experience
          in: query
          description: doctor wokrk_experience
          required: true
          schema:
            type: string
        - name: schedule
          in: query
          description: doctor schedule
          required: true
          schema:
            type: object
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                properties:
                  meta:
                    $ref: "#/components/schemas/SuccessResponse"
                  data:
                    $ref: "#/components/schemas/DoctorResponses"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FailResponse400"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FailResponse403"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FailResponse500"
  /doctor/{uuid}:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - Doctor
      summary: Get Doctor by uuid
      parameters:
        - in: path
          name: uuid
          schema:
            type: string
          required: true
          description: Doctor uuid
        - in: query
          name: doctor
          required: true
          schema:
            $ref: "#/components/schemas/Doctor"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                properties:
                  meta:
                    $ref: "#/components/schemas/SuccessResponse"
                  data:
                    $ref: "#/components/schemas/DoctorResponses"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FailResponse400"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FailResponse500"
  /doctor/?nik={nik}:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - Doctor
      summary: Get Doctor by uuid
      parameters:
        - in: path
          name: uuid
          schema:
            type: string
          required: true
          description: Doctor uuid
        - in: query
          name: doctor
          required: true
          schema:
            $ref: "#/components/schemas/Doctor"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                properties:
                  meta:
                    $ref: "#/components/schemas/SuccessResponse"
                  data:
                    $ref: "#/components/schemas/DoctorResponses"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FailResponse400"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FailResponse500"
  /doctor/?name={name}:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - Doctor
      summary: Get Doctor by uuid
      parameters:
        - in: path
          name: uuid
          schema:
            type: string
          required: true
          description: Doctor uuid
        - in: query
          name: doctor
          required: true
          schema:
            $ref: "#/components/schemas/Doctor"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                properties:
                  meta:
                    $ref: "#/components/schemas/SuccessResponse"
                  data:
                    $ref: "#/components/schemas/DoctorResponses"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FailResponse400"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FailResponse500"
...
  /nurse/login:
    post:
      tags:
        - Nurse
      summary: Nurse Login
      operationId: loginNurse
      parameters:
        - name: email
          in: query
          description: The email Nurse for login
          required: true
          schema:
            type: string
        - name: password
          in: query
          description: The password Nurse for login
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: "#/components/schemas/Login"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FailResponse400"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FailResponse500"
  /nurse/register:
    post:
      tags:
        - Nurse
      summary: Register new Nurse
      parameters:
        - name: name
          in: query
          description: nurse Name
          required: true
          schema:
            type: string
        - name: nik
          in: query
          description: nurse Nik
          required: true
          schema:
            type: string
        - name: password
          in: query
          description: nurse Password
          required: true
          schema:
            type: string
        - name: dob
          in: query
          description: nurse Date of Birth
          required: true
          schema:
            type: string
        - name: sex
          in: query
          description: nurse Sex
          required: true
          schema:
            type: string
        - name: email
          in: query
          description: nurse Email
          required: true
          schema:
            type: string
        - name: contact
          in: query
          description: nurse Contact
          required: true
          schema:
            type: string
        - name: wokrk_experience
          in: query
          description: nurse wokrk_experience
          required: true
          schema:
            type: string
        - name: schedule
          in: query
          description: nurse schedule
          required: true
          schema:
            type: string
        - name: work_hour
          in: query
          description: nurse work_hour
          required: true
          schema:
            type: string

      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                properties:
                  meta:
                    $ref: "#/components/schemas/SuccessResponse"
                  data:
                    $ref: "#/components/schemas/NurseResponses"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FailResponse400"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FailResponse500"
...
  /nurse:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - Nurse
      summary: Get all Nurse
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                properties:
                  meta:
                    $ref: "#/components/schemas/SuccessResponse"
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/NurseResponses"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FailResponse404"
    delete:
      security:
        - bearerAuth: [ ]
      tags:
        - Nurse
      summary: Delete a specified nurse
      parameters:
        - in: path
          name: uuid
          schema:
            type: string
          description: uuid of nurse to be deleted
          required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                properties:
                  meta:
                    $ref: "#/components/schemas/SuccessResponse"
                  data:
                    type: string
                    example: nurse was Deleted
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FailResponse400"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FailResponse403"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FailResponse500"
    put:
      security:
        - bearerAuth: [ ]
      tags:
        - Nurse
      summary: Update nurse
      parameters:
        - name: name
          in: query
          description: nurse Name
          required: true
          schema:
            type: string
        - name: password
          in: query
          description: nurse Password
          required: true
          schema:
            type: string
        - name: dob
          in: query
          description: nurse Date of Birth
          required: true
          schema:
            type: string
        - name: sex
          in: query
          description: nurse Sex
          required: true
          schema:
            type: string
        - name: email
          in: query
          description: nurse Email
          required: true
          schema:
            type: string
        - name: contact
          in: query
          description: nurse Contact
          required: true
          schema:
            type: string
        - name: wokrk_experience
          in: query
          description: nurse wokrk_experience
          required: true
          schema:
            type: string
        - name: schedule
          in: query
          description: nurse schedule
          required: true
          schema:
            type: object
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                properties:
                  meta:
                    $ref: "#/components/schemas/SuccessResponse"
                  data:
                    $ref: "#/components/schemas/NurseResponses"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FailResponse400"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FailResponse403"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FailResponse500"
  /nurse/{uuid}:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - Nurse
      summary: Get Nurse by uuid
      parameters:
        - in: path
          name: uuid
          schema:
            type: string
          required: true
          description: Nurse uuid
        - in: query
          name: Nurse
          required: true
          schema:
            $ref: "#/components/schemas/Nurse"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                properties:
                  meta:
                    $ref: "#/components/schemas/SuccessResponse"
                  data:
                    $ref: "#/components/schemas/NurseResponses"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FailResponse400"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FailResponse500"
  /nurse?nik={nik}:
    post:
      security:
        - bearerAuth: [ ]
      tags:
        - Nurse
      summary: Get Nurse by uuid
      parameters:
        - in: path
          name: uuid
          schema:
            type: string
          required: true
          description: Nurse uuid
        - in: query
          name: Nurse
          required: true
          schema:
            $ref: "#/components/schemas/Nurse"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                properties:
                  meta:
                    $ref: "#/components/schemas/SuccessResponse"
                  data:
                    $ref: "#/components/schemas/NurseResponses"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FailResponse400"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FailResponse500"
  /nurse?name={name}:
    post:
      security:
        - bearerAuth: [ ]
      tags:
        - Nurse
      summary: Get Nurse by uuid
      parameters:
        - in: path
          name: uuid
          schema:
            type: string
          required: true
          description: Nurse uuid
        - in: query
          name: Nurse
          required: true
          schema:
            $ref: "#/components/schemas/Nurse"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                properties:
                  meta:
                    $ref: "#/components/schemas/SuccessResponse"
                  data:
                    $ref: "#/components/schemas/NurseResponses"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FailResponse400"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FailResponse500"
...
  /patient?uuid={uuid}:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - Patient
      summary: Get Patient by uuid
      parameters:
        - in: path
          name: uuid
          schema:
            type: string
          required: true
          description: Patient uuid
        - in: query
          name: Patient
          required: true
          schema:
            $ref: "#/components/schemas/Patient"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                properties:
                  meta:
                    $ref: "#/components/schemas/SuccessResponse"
                  data:
                    $ref: "#/components/schemas/PatientResponses"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FailResponse400"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FailResponse500"
...
  /WorkDay/{day}:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - Work Day
      summary: Get Work Day by day
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                properties:
                  meta:
                    $ref: "#/components/schemas/SuccessResponse"
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/WorkDayResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FailResponse404"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Login:
      type: object
      properties:
        meta:
          type: object
          properties:
            status:
              type: integer
              example: 200
            message:
              type: string
              example: successful to login
            token:
              type: string
              example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJhdXRob3JpemVkIjp0cnVlLCJleHAiOjE2MzIyMTU5ODUsInVzZXJpZCI6MTF9.N72CvB3ffoZKvntUC4MAtDPpzQTQjIfzvnNSTNBuC7A
...
    SuccessResponse:
      type: object
      properties:
        meta:
          type: object
          properties:
            status:
              type: integer
              example: 200
            message:
              type: string
              example: Success
    FailResponse400:
      type: object
      properties:
        meta:
          type: object
          properties:
            status:
              type: integer
              example: 400
            message:
              type: string
              example: Error
            messages:
              type: array
              items:
                example: Error message goes here
        data:
          type: string
          example: null
...
    FailResponse500:
      type: object
      properties:
        meta:
          type: object
          properties:
            status:
              type: integer
              example: 500
            message:
              type: string
              example: Error
            messages:
              type: array
              items:
                example: Error message goes here
        data:
          type: string
          example: null